SDC-9275. Directory Origin fails with read order set to Timestamp
* the issue is the origin does not pick up the file from the initial start when the first file to process is empty. After the initial load, it starts to pick up newly created/moved files in the directory.
* when we call Path(""), it returns the SDC_HOME_DIRECTORY

Reviewed-on: https://review.streamsets.net/15156
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ basic-lib/src/main/java/com/streamsets/pipeline/lib/dirspooler/LocalFileSystem.java
@@ -65,10 +65,16 @@ public class LocalFileSystem implements WrappedFileSystem {
   }
 
   public boolean exists(WrappedFile filePath) {
+    if (filePath == null) {
+      return false;
+    }
     return Files.exists(Paths.get(filePath.getAbsolutePath()));
   }
 
   public void delete(WrappedFile filePath) throws IOException {
+    if (filePath == null) {
+      return;
+    }
     Files.delete(Paths.get(filePath.getAbsolutePath()));
   }
 
@@ -77,10 +83,16 @@ public class LocalFileSystem implements WrappedFileSystem {
   }
 
   public long getLastModifiedTime(WrappedFile filePath) throws IOException {
+    if (filePath == null) {
+      return -1;
+    }
     return getLastModifiedTime(Paths.get(filePath.getAbsolutePath()));
   }
 
   public long getChangedTime(WrappedFile filePath) throws IOException {
+    if (filePath == null) {
+      return -1;
+    }
     return getChangedTime(Paths.get(filePath.getAbsolutePath()));
   }
 
@@ -93,6 +105,9 @@ public class LocalFileSystem implements WrappedFileSystem {
   }
 
   public boolean isDirectory(WrappedFile filePath) {
+    if (filePath == null) {
+      return false;
+    }
     return Files.isDirectory(Paths.get(filePath.getAbsolutePath()));
   }
